+++ sql-plugin/src/main/321db/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -129,15 +129,15 @@ object SparkShimImpl extends Spark321PlusShims with Spark320until340Shims with S
             val convertBroadcast = (bc: SubqueryBroadcastExec) => {
               val meta = GpuOverrides.wrapAndTagPlan(bc, conf)
               meta.tagForExplain()
-              val converted = meta.convertIfNeeded() 
+              val converted = meta.convertIfNeeded()
               // Because the PlanSubqueries rule is not called (and does not work as expected),
-              // we might actually have to fully convert the subquery plan as the plugin would 
-              // intend (in this case calling GpuTransitionOverrides to insert GpuCoalesceBatches, 
+              // we might actually have to fully convert the subquery plan as the plugin would
+              // intend (in this case calling GpuTransitionOverrides to insert GpuCoalesceBatches,
               // etc.) to match the other side of the join to reuse the BroadcastExchange.
               // This happens when SubqueryBroadcast has the original (Gpu)BroadcastExchangeExec
               converted match {
                 case e: GpuSubqueryBroadcastExec => e.child match {
-                  // If the GpuBroadcastExchange is here, then we will need to run the transition 
+                  // If the GpuBroadcastExchange is here, then we will need to run the transition
                   // overrides here
                   case _: GpuBroadcastExchangeExec =>
                     var updated = ApplyColumnarRulesAndInsertTransitions(Seq(), true)
@@ -153,7 +153,7 @@ object SparkShimImpl extends Spark321PlusShims with Spark320until340Shims with S
                     }
                   // Otherwise, if this SubqueryBroadcast is using a ReusedExchange, then we don't
                   // do anything further
-                  case _: ReusedExchangeExec => 
+                  case _: ReusedExchangeExec =>
                     converted.asInstanceOf[BaseSubqueryExec]
                 }
                 case _ =>
@@ -262,6 +262,3 @@ object SparkShimImpl extends Spark321PlusShims with Spark320until340Shims with S
   //   case BroadcastQueryStageExec(_, e: ReusedExchangeExec, _, _) => e
   // }
 }
-
-// Fallback to the default definition of `deterministic`
-trait GpuDeterministicFirstLastCollectShim extends Expression
