+++ sql-plugin/src/main/330/scala/com/nvidia/spark/rapids/SparkShims.scala
@@ -18,7 +18,8 @@ package com.nvidia.spark.rapids.shims
 
 import com.nvidia.spark.rapids._
 
-object SparkShimImpl extends Spark330PlusNonDBShims
+object SparkShimImpl extends Spark330PlusShims
+  with Spark330PlusNonDBShims
   with Spark320until340Shims
   with Spark320PlusNonDBShims {
   override def getSparkShimVersion: ShimVersion = ShimLoader.getShimVersion
