+++ sql-plugin/src/main/330until340-nondb/scala/org/apache/spark/sql/rapids/shims/RapidsErrorUtils.scala
